% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin-js.R
\name{odin_js}
\alias{odin_js}
\alias{odin_js_}
\title{Create a javascript odin model}
\usage{
odin_js(x)

odin_js_(x)
}
\arguments{
\item{x}{Either the name of a file to read, a text string (if
length is greater than 1 elements will be joined with newlines)
or an expression.}
}
\description{
Generate a javascript odin model from a file, text string or
expression.
}
